+++ app/controllers/local_authorities_controller.rb
@@ -7,6 +7,22 @@ class LocalAuthoritiesController < ApplicationController
     @local_authority = LocalAuthority.find(params[:id])
   end
 
+  def new
+    @local_authority = LocalAuthority.new
+  end
+
+  def create
+    @local_authority = LocalAuthority.new(local_authority_params)
+
+    if @local_authority.valid?
+      @local_authority.save
+
+      redirect_to local_authority_path(@local_authority), notice: I18n.t("local_authority.create.success")
+    else
+      render :new
+    end
+  end
+
   def edit
     @local_authority = LocalAuthority.find(params[:id])
   end
@@ -26,6 +42,6 @@ class LocalAuthoritiesController < ApplicationController
 
   private def local_authority_params
     params.require(:local_authority)
-      .permit(:code, :address_1, :address_2, :address_3, :address_town, :address_county, :address_postcode)
+      .permit(:name, :code, :address_1, :address_2, :address_3, :address_town, :address_county, :address_postcode)
   end
 end