+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added a new list of all projects with Academy URNs
 - Display local authorities and their details
 - Allow local authority records to be edited
+- Allow new local authority records to be added
 
 ### Changed